LayoutTests/fast/encoding/charset-replacement.html

<!DOCTYPE html>
<html>
<head>
    <meta charset=rEpLaCeMeNt>
    <script>
    if (window.testRunner)
        testRunner.dumpAsText();
    alert(document.characterSet);
    </script>
</head>
<body>
    Test PASSED if the encoding of this document is the default encoding.<br>
    Test FAILED if you see a U+FFFD character in a dumped render tree.<br>
</body>
</html>
